refactor, bug fix
