+++ internal/bot/controller/controller.go
@@ -103,15 +103,15 @@ func (c *Controller) SaveUsers(ctx context.Context, users []*user_model.User) {
 			errors = append(errors, err)
 		}
 
-		err = c.reminderSrv.CreateScheduler(u.TGID, loc)
+		err = c.reminderSrv.CreateScheduler(ctx, u.TGID, loc, c.SendReminder)
 		if err != nil {
 			errors = append(errors, err)
 		}
 
-		err = c.reminderSrv.StartAllJobs(ctx, u.TGID, loc, c.SendReminder)
-		if err != nil {
-			errors = append(errors, err)
-		}
+		// err = c.reminderSrv.StartAllJobs(ctx, u.TGID, loc, c.SendReminder)
+		// if err != nil {
+		// 	errors = append(errors, err)
+		// }
 	}
 
 	if len(errors) > 0 {
@@ -160,5 +160,5 @@ func (c *Controller) saveUser(ctx context.Context, tgID int64) error {
 		return err
 	}
 
-	return c.reminderSrv.CreateScheduler(tgID, loc)
+	return c.reminderSrv.CreateScheduler(ctx, tgID, loc, c.SendReminder)
 }
